Fixed login and registration authorization, removed unnecessary things and refactored.
