+++ src/main/java/com/backend/appbackend/security/JWTAuthenticationFilter.java
@@ -1,9 +1,8 @@
 package com.backend.appbackend.security;
 
 import com.auth0.jwt.JWT;
-import com.backend.appbackend.user.entity.User;
+import com.backend.appbackend.user.User;
 import com.fasterxml.jackson.databind.ObjectMapper;
-
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
