+++ src/main/java/com/backend/appbackend/security/JWTAuthorizationFilter.java
@@ -14,9 +14,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.ArrayList;
 
-import static com.backend.appbackend.security.SecurityConstants.HEADER_STRING;
-import static com.backend.appbackend.security.SecurityConstants.TOKEN_PREFIX;
-import static com.backend.appbackend.security.SecurityConstants.SECRET;
+import static com.backend.appbackend.security.SecurityConstants.*;
 
 public class JWTAuthorizationFilter extends BasicAuthenticationFilter {
