+++ src/main/java/com/backend/appbackend/security/SecurityConstants.java
@@ -6,4 +6,5 @@ public class SecurityConstants {
     public static final String TOKEN_PREFIX = "Bearer ";
     public static final String HEADER_STRING = "Authorization";
     public static final String SIGN_UP_URL = "/sign-up";
+    public static final String VIEW_JOBS_URL = "/jobs/all";
 }
