+++ src/main/java/com/backend/appbackend/user/User.java
@@ -1,4 +1,4 @@
-package com.backend.appbackend.user.entity;
+package com.backend.appbackend.user;
 
 import javax.persistence.Id;
