+++ src/main/java/com/backend/appbackend/user/UserRepository.java
@@ -1,9 +1,10 @@
-package com.backend.appbackend.user.repository;
+package com.backend.appbackend.user;
 
-import com.backend.appbackend.user.entity.User;
+import com.backend.appbackend.user.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface UserRepository extends MongoRepository<User, String> {
     User findUserById(String id);
+
     User findUserByEmail(String email);
 }
