+++ src/main/java/com/backend/appbackend/user/UserServiceImpl.java
@@ -1,8 +1,9 @@
-package com.backend.appbackend.user.service;
+package com.backend.appbackend.user;
 
-import com.backend.appbackend.user.entity.User;
-import com.backend.appbackend.user.exception.UserNotFoundException;
-import com.backend.appbackend.user.repository.UserRepository;
+import com.backend.appbackend.user.User;
+import com.backend.appbackend.user.UserNotFoundException;
+import com.backend.appbackend.user.UserRepository;
+import com.backend.appbackend.user.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
